Corrected outlier delimiters to match augur parse output
The augur parse subcommand replaces hyphens with underscores in strain names in
both sequence headers and metadata. As a result, outlier names with hyphens are
not properly filtered. This commit renames outliers to match the augur parse
output format and also adds one strain from Indiana that is likely to be swine
flu based on BLAST alignment.