Align builds with seasonal flu and increase available data
Specifically, this commit increases the number of viruses sampled per month to
get approximately ~9,000 total viruses in the tree and it downloads titer data
from the "base" titer database in addition to the CDC titers resulting in a
doubling of available measurements.

Updates to the Snakefile rules mostly align this build with the seasonal flu
repository while updates to the config JSON include titer tree model values in
the final output table.