Fix logic for binning nodes in tree by seasons
Prior to this commit, nodes were binned by season using the last date of the
season. This led to unexpected behavior of a strain's LBI or distance being
reported as belonging to the season _after_ the strain was sampled. This commit
fixes that behavior by binning seasons in unambiguous pairs of dates for the
season start and end, respectively.

As part of this change, this commit also speeds up the assignment of nodes to
seasons by using the numpy searchsorted function and cleans up some boilerplate
code by using a defaultdict. The two scripts that rely on the
`get_nodes_by_seasons` function had to be updated to reflect the new season key
for nodes, but this also led to simpler code in those scripts.