Force augur to include 'date' field in auspice JSONs
We aggregate multiple node data JSONs and tree information into auspice JSONs
and then convert these into tables downstream. However, augur export v2 excludes
some fields like 'date' from export by default [1]. This commit forces the
'date' field to be included as a color-by field. This feels like a hack and may
not be in the spirit of the augur export v2 logic, but it beats adding a new
step to our workflow with a custom script that basically does the same thing as
augur export without excluding fields.

[1] https://github.com/nextstrain/augur/blob/98c0ea34caf341e1b1ae9ffd335f0f9473f5afbd/augur/export_v2.py#L643-L668