Calculate pairwise titer distances to current vaccine per season
Adds support for calculating pairwise titer distances to the current vaccine for
all strains in a given season. This commit adds a new `--vaccines` input
argument to the seasonal distances script. When users provide this argument, the
script calculates pairwise distances from each tip to the corresponding season's
vaccine instead of the tip's most recent ancestor in the previous season.