Calculate titer tree model distances for seasons
Add a script and rule to calculate mean distances between seasons with
the titer tree model. This is an altered version of the original mean
seasonal distance script that needs to find the path between pairs of
strains in a tree and sum the weight of titer drops on that path. This
implementation is a copy of code from the flu-forecasting project [1].

[1] https://github.com/blab/flu-forecasting/blob/723c515/scripts/pairwise_titer_tree_distances.py