Exclude egg-passaged reference strains from analysis
We do not want to include _any_ egg-passaged strains in the analysis, but the
way the augur filter command was setup in this workflow, we excluded most
egg-passaged strains and then force-included those in the list of reference
strains. This change drops those egg-passaged reference strains from the
metadata, too, so they won't make it into the trees.

Closes #5