Build NA trees without linking to HA trees
Builds NA trees and calculates distances for these trees without requiring HA
and NA trees to have the same samples. Relaxing this original constraint in the
workflow allows HA trees to include more samples at earlier timepoints when
sequencing of NA was not as routine. The main change here is to parameterize the
"select strains" rule by segment instead of passing both segments to this
script. The result is a list of strains per segment that the rest of the
workflow can source downstream without many other changes. The primary change
required is to avoid running titer models on the NA trees.

Closes #7