Fix a bug in tree to table conversion
Fixes a bug caused when requested columns in the tree to table conversion final
output don't exist in the given data frame. This bug existed for a while as a
deprecation warning with earlier versions of pandas, but with the latest
versions of pandas indexing the `.loc` operator with columns that don't exist
produces a `KeyError`.